adding weapon damage to combat action hp change function init and tests

+++ common/src/combat/ability_handlers/split_combat_action_hp_change_by_number_of_targets.rs
@@ -1,27 +1,24 @@
 use crate::app_consts::MULTI_TARGET_HP_CHANGE_BONUS;
 
 pub fn split_combat_action_hp_change_by_number_of_targets(
-    min_hp_change: u16,
-    max_hp_change: u16,
-    num_targets: u8,
+    min_hp_change: f32,
+    max_hp_change: f32,
+    num_targets: f32,
 ) -> (f32, f32) {
-    let multi_target_bonus = 1.0 + (num_targets as f32 - 1.0) * MULTI_TARGET_HP_CHANGE_BONUS;
+    let multi_target_bonus = 1.0 + (num_targets - 1.0) * MULTI_TARGET_HP_CHANGE_BONUS;
 
-    let min_hp_change = min_hp_change as f32 * multi_target_bonus;
-    let max_hp_change = max_hp_change as f32 * multi_target_bonus;
+    let min_hp_change = min_hp_change * multi_target_bonus;
+    let max_hp_change = max_hp_change * multi_target_bonus;
     // split hp_change between all targets
-    (
-        min_hp_change / num_targets as f32,
-        max_hp_change / num_targets as f32,
-    )
+    (min_hp_change / num_targets, max_hp_change / num_targets)
 }
 
 #[cfg(test)]
 #[test]
 fn split_combat_action_hp_change_by_number_of_targets_test() {
-    let min_hp_change: u16 = 30;
-    let max_hp_change: u16 = 300;
-    let num_targets = 3;
+    let min_hp_change: f32 = 30.0;
+    let max_hp_change: f32 = 300.0;
+    let num_targets = 3.0;
     let (new_min, new_max) = split_combat_action_hp_change_by_number_of_targets(
         min_hp_change,
         max_hp_change,
@@ -29,9 +26,9 @@ fn split_combat_action_hp_change_by_number_of_targets_test() {
     );
     assert_eq!(new_min, 13.0);
     assert_eq!(new_max, 130.0);
-    let min_hp_change: u16 = 20;
-    let max_hp_change: u16 = 200;
-    let num_targets = 2;
+    let min_hp_change: f32 = 20.0;
+    let max_hp_change: f32 = 200.0;
+    let num_targets = 2.0;
     let (new_min, new_max) = split_combat_action_hp_change_by_number_of_targets(
         min_hp_change,
         max_hp_change,
